+++ R/combine_plots.R
@@ -38,7 +38,7 @@
 #' @param sub.vpadding Vertical padding. The total vertical space added to the
 #'   label, given in grid units. By default, this is added equally above and
 #'   below the label. However, by changing the y and vjust parameters, this can
-#'   be changed (Default: `grid::unit(1, "lines")`).
+#'   be changed (Default: `ggplot2::unit(1, "lines")`).
 #' @param sub.fontface The font face (`"plain"` (default), `"bold"`, `"italic"`,
 #'   `"bold.italic"`) for the annotation label.
 #' @param sub.angle Angle at which annotation label is to be drawn (Default:
@@ -52,7 +52,6 @@
 #'   combining (plot, caption).
 #'
 #' @importFrom cowplot plot_grid add_sub ggdraw draw_label
-#' @importFrom grid unit
 #'
 #' @references
 #' \url{https://indrajeetpatil.github.io/ggstatsplot/articles/web_only/combine_plots.html}
@@ -113,7 +112,7 @@ combine_plots <- function(...,
                           sub.fontface = "plain",
                           sub.x = 0.5,
                           sub.y = 0.5,
-                          sub.vpadding = grid::unit(1, "lines"),
+                          sub.vpadding = ggplot2::unit(1, "lines"),
                           sub.angle = 0,
                           sub.lineheight = 0.9,
                           title.rel.heights = c(0.1, 1.2),